+++ platform/nxp/lpc17xx/startup.c
@@ -4,6 +4,7 @@
  * Project is distributed under the terms of the GNU General Public License v3.0
  */
 
+#include <stddef.h>
 #include <xcore/bits.h>
 #include <halm/platform/nxp/lpc17xx/system.h>
 #include <halm/platform/nxp/lpc17xx/system_defs.h>
@@ -29,6 +30,6 @@ void platformStartup(void)
    * Other peripherals are disabled by default.
    */
 
-  for (unsigned int index = 0; index < ARRAY_SIZE(blocksToDisable); ++index)
+  for (size_t index = 0; index < ARRAY_SIZE(blocksToDisable); ++index)
     sysPowerDisable(blocksToDisable[index]);
 }
